Handle snaktype: «somevalue».
From http://www.wikidata.org/wiki/Wikidata:Glossary/en#Property

« Unknown value is a marker for when there is some value but the exact
value is not known for the property. "Some value" means that there is
nothing known for the value except that it should exist and not imply a
negation of the claim (example: if the date of a human's death is
completely unknown the item would receive this marker for date of death
(P570), denoting that the human is, in fact, dead — however, with the
date of death being unknown). »